Let WorldObjects have "height."
* Height, being an extra value offsetting the object's Y position
without actually modifying their Y position, therefore giving the object
a sense of height. This means if this object is in front of something
and it has height, it will still be in front of that something despite
having a Y value that might say otherwise, based on internal rendering
code.