Implement "thresholding" for ImageLayers.
! A threshold value resolves the issue for images that're not shaped as
a perfect rectangle. So long as the map defines collision boundaries for
the image correctly, WorldObjects on the same layer as the image will
not jump behind the image until the object's position is at (threshold *
imageHeight) pixels more above the image's original position (the
original position being at the bottom-left of the image).